added Sonic Rush Adventure model
finally, I was able to add Sonic's model from Sonic Rush Adventure for Modern Sonic! this is what I was trying to do from the start, but the only rip of the model on the internet was deleted sometime in the last couple years, and the original ripper forgot how they did it. sooo I had to hop around between various forums to figure out how to get apicula (the DS model ripping program for .nsbmd files, which most DS games including the Sonic Rush games use) to export animations in a way that blender could read (I just had to add -fglb as a parameter the whole time oops) so hours of troubleshooting later... Sonic Rush!